Player from AR-P

+++ player2_191/components/include/ring.h
@@ -25,6 +25,7 @@ Author :           Nick
 Definition of the pure virtual class defining the interface to a simple ring
 storage device.
 
+
 Date        Modification                                    Name
 ----        ------------                                    --------
 11-Dec-03   Created                                         Nick
@@ -36,34 +37,34 @@ Date        Modification                                    Name
 
 #include "osinline.h"
 
-typedef enum
+typedef enum 
 {
-	RingNoError     = 0,
-	RingNoMemory,
-	RingTooManyEntries,
-	RingNothingToGet
+    RingNoError		= 0,
+    RingNoMemory,
+    RingTooManyEntries,
+    RingNothingToGet
 } RingStatus_t;
 
 //
 
-#define RING_NONE_BLOCKING  0
+#define	RING_NONE_BLOCKING	0
 
 //
 
 class Ring_c
 {
-	public:
+public:
 
-		RingStatus_t    InitializationStatus;
+    RingStatus_t	InitializationStatus;
 
-		virtual ~Ring_c(void) {};
+    virtual ~Ring_c( void ) {};
 
-		virtual RingStatus_t Insert(uintptr_t Value) = 0;
-		virtual RingStatus_t Extract(uintptr_t *Value,
-									 unsigned int   BlockingPeriod = RING_NONE_BLOCKING) = 0;
-		virtual RingStatus_t Flush(void) = 0;
-		virtual bool         NonEmpty(void) = 0;
+    virtual RingStatus_t Insert(  unsigned int	 Value ) = 0;
+    virtual RingStatus_t Extract( unsigned int  *Value,
+                                  unsigned int   BlockingPeriod = RING_NONE_BLOCKING ) = 0;
+    virtual RingStatus_t Flush( void ) = 0;
+    virtual bool         NonEmpty( void ) = 0;
 };
 
-typedef class Ring_c    *Ring_t;
+typedef class Ring_c	*Ring_t;
 #endif
